src/worklog/requirements.txt
python-telegram-bot>=21.0
openai>=1.0
pymongo>=4.0
pytz>=2024.1
python-dotenv>=1.0
